Organize session data-viz package
