+++ frontend/src/components/pages/ComponentLibrary.tsx
@@ -6,11 +6,11 @@ import { Grade } from "../../APIClients/types/UserClientTypes";
 import type { ClassroomForm } from "../../types/ClassroomTypes";
 import StudentDashboardHeader from "../assessments/assessment-creation/StudentDashboardHeader";
 import ClassroomCard from "../classrooms/ClassroomCard";
-import CorrectedMultipleChoice from "../sessions/results/CorrectedMultipleChoice";
-import CorrectedMultiSelect from "../sessions/results/CorrectedMultiSelect";
-import CorrectedShortAnswer from "../sessions/results/CorrectedShortAnswer";
-import StatisticCard from "../sessions/results/StatisticCard";
-import StudentList from "../sessions/results/StudentList";
+import CorrectedMultipleChoice from "../sessions/data-visualization/corrected-question-elements/CorrectedMultipleChoice";
+import CorrectedMultiSelect from "../sessions/data-visualization/corrected-question-elements/CorrectedMultiSelect";
+import CorrectedShortAnswer from "../sessions/data-visualization/corrected-question-elements/CorrectedShortAnswer";
+import StatisticCard from "../sessions/data-visualization/StatisticCard";
+import StudentList from "../sessions/data-visualization/StudentList";
 import AddClassroomModal from "../user-management/student/AddClassroomModal";
 import AddStudentModal from "../user-management/student/AddStudentModal";
