+++ frontend/src/components/sessions/data-visualization/StatisticCard.tsx
