+++ frontend/src/components/sessions/data-visualization/StudentList/StudentList.tsx
@@ -1,61 +1,12 @@
 import React, { useMemo, useState } from "react";
-import {
-  Box,
-  Button,
-  Divider,
-  Flex,
-  ListItem,
-  OrderedList,
-  Spacer,
-  VStack,
-} from "@chakra-ui/react";
+import { Divider, Flex, OrderedList, VStack } from "@chakra-ui/react";
 
-import type { StudentResponse } from "../../../APIClients/types/ClassClientTypes";
-import { includesIgnoreCase } from "../../../utils/GeneralUtils";
-import SearchBar from "../../common/table/SearchBar";
-import SortMenu from "../../common/table/SortMenu";
+import type { StudentResponse } from "../../../../APIClients/types/ClassClientTypes";
+import { includesIgnoreCase } from "../../../../utils/GeneralUtils";
+import SearchBar from "../../../common/table/SearchBar";
+import SortMenu from "../../../common/table/SortMenu";
 
-type StudentListItemProps = {
-  firstName: string;
-  lastName: string;
-  isViewed: boolean;
-  isSelected: boolean;
-  onClick: () => void;
-};
-
-const StudentListItem = ({
-  firstName,
-  lastName,
-  isViewed,
-  isSelected,
-  onClick,
-}: StudentListItemProps) => {
-  return (
-    <ListItem>
-      <Button
-        _hover={{ bg: "grey.100" }}
-        bg="grey.50"
-        borderRadius={8}
-        display="flex"
-        onClick={onClick}
-        p={4}
-        variant="tertiary"
-        w="100%"
-      >
-        <Box
-          color={isSelected ? "blue.300" : "grey.400"}
-          fontWeight={isSelected ? "700" : "400"}
-        >{`${firstName} ${lastName}`}</Box>
-        <Spacer />
-        {isViewed && (
-          <Box color="grey.200" fontWeight="400">
-            Viewed
-          </Box>
-        )}
-      </Button>
-    </ListItem>
-  );
-};
+import StudentListItem from "./StudentListItem";
 
 type Student = StudentResponse & {
   isViewed: boolean;
