+++ frontend/src/components/sessions/data-visualization/StudentList/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./StudentList";
