+++ frontend/src/components/sessions/data-visualization/corrected-question-elements/CorrectedInputWrapper.tsx
