+++ frontend/src/components/sessions/data-visualization/corrected-question-elements/CorrectedMultiOptionElements.tsx
